+++ src/App.js
@@ -1,38 +1,8 @@
 import { BrowserRouter, Link, Outlet, Route, Routes } from "react-router-dom";
-import { createGlobalStyle } from "styled-components";
-import reset from "styled-reset";
-import UploadProduct from "./pages/UploadProduct";
-import ProductDetail from "./pages/ProductDetail";
-import ProductList from "./pages/ProductList";
-
-const GlobalStyle = createGlobalStyle`
-  ${reset}
-  :root{
-    --main-color: rgba(5, 139, 46, 1);
-  }
-  #root{
-    width: min(80vw, 500px);
-    margin: 0 auto;
-  }
-  .a11y-hidden {
-    clip: rect(1px, 1px, 1px, 1px);
-    clip-path: inset(50%);
-    width: 1px;
-    height: 1px;
-    margin: -1px;
-    overflow: hidden;
-    padding: 0;
-    position: absolute;
-  }
-  img{
-    max-width: 100%;
-    vertical-align: top;
-  }
-  a{
-    text-decoration: none;
-    color: initial;
-  }
-`;
+import UploadProduct from "./pages/UploadProduct/UploadProduct";
+import ProductDetail from "./pages/ProductDetail/ProductDetail";
+import ProductList from "./pages/ProductList/ProductList";
+import GlobalStyle from "./styles/GlobalStyle";
 
 function App() {
   return (
