+++ src/pages/ProductDetail/ProductDetailStyle.jsx
@@ -1,0 +1,9 @@
+import styled from "styled-components";
+
+const ProductPage = styled.main``;
+
+const ProductImage = styled.img`
+  width: 100%;
+`;
+
+export { ProductImage, ProductPage };
