+++ src/pages/ProductList/ProductListStyle.jsx
@@ -1,6 +1,4 @@
 import styled from "styled-components";
-import Products from "../components/Products";
-import { useRef, useState } from "react";
 
 const Section = styled.section`
   margin-left: 20px;
@@ -40,21 +38,4 @@ const Ul = styled.ul`
   }
 `;
 
-export default function ProductList() {
-  const userAccountName = "testtestabc";
-  const slideUl = useRef(null);
-  const [mouseX, setmouseX] = useState(0);
-
-  function test(e) {
-    setmouseX(e.clientX);
-  }
-
-  return (
-    <Section>
-      <H2 className="product-section-title">판매 중인 상품</H2>
-      <Ul ref={slideUl} onDrag={test} style={{ transform: `translateX: ${mouseX}px` }}>
-        <Products userAccountName={userAccountName} />
-      </Ul>
-    </Section>
-  );
-}
+export { Section, H2, Ul };
